已入库数量返回值优化。

+++ src/main/java/com/ev/scm/dao/StockInItemDao.java
@@ -4,6 +4,7 @@ package com.ev.scm.dao;
 import com.ev.scm.domain.StockInItemDO;
 import org.apache.ibatis.annotations.Mapper;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -60,6 +61,6 @@ public interface StockInItemDao {
 
 	List<Map<String, Object>>  getItemDate(Map<String, Object> map);
 
-	int getInCountOfContract(Map<String, Object> map);
+	BigDecimal getInCountOfContract(Map<String, Object> map);
 
 }
